+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Apply all modifiers preset modifiers before user modifiers from `with`.
+
 ## [0.20.0] - 2025-11-15
 
 ### Added