+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.20.1] - 2025-12-29
+
 ### Changed
 
 - All preset modifiers are now applied before user modifiers from `with` to improve intuitiveness of construction.
@@ -491,7 +493,8 @@ This update features a big rewrite into a component-based API. The core concepts
 
 Initial release.
 
-[unreleased]: https://github.com/simgine/bevy_replicon/compare/v0.20.0...HEAD
+[unreleased]: https://github.com/simgine/bevy_replicon/compare/v0.20.1...HEAD
+[0.20.1]: https://github.com/simgine/bevy_replicon/compare/v0.20.0...v0.20.1
 [0.20.0]: https://github.com/simgine/bevy_replicon/compare/v0.19.3...v0.20.0
 [0.19.3]: https://github.com/simgine/bevy_replicon/compare/v0.19.2...v0.19.3
 [0.19.2]: https://github.com/simgine/bevy_replicon/compare/v0.19.1...v0.19.2